error fixed remove log_files directory

+++ python-code/hiphive-scripts/get-force-constants.py
@@ -25,7 +25,8 @@ atoms_ideal = read('supercell_ideal.extxyz') # ideal supercell
 # create the cluster space for the defined cutoffs
 cs = ClusterSpace(prim, cutoffs)
 
-shutil.rmtree('log_files')
+if os.path.exists('log_files'):
+    shutil.rmtree('log_files')
 os.mkdir('log_files')
 
 with open('log_files/1-cluster-space', 'w') as f: # save the information in a file
